requirements.txt
ragas>=0.1.0
pymongo>=4.6.0
python-dotenv>=1.0.0
datasets>=2.18.0
openai>=1.0.0
